Fix printing of watched filename when watching a file (#157)
When watching a directory, %f should print the file name and %w should
print the watched directory.

When watching a single file, %f should print an empty string and %w
should print the watched filename.

Commit 0971a97 ("Fanotify support (#134)") changed that for fanotify
because watching a single file or its directory produces the same
fanotify event with non-empty filename.

Revert that change for inotify to fix the regression, but leave it for
fanotify. Update man page to reflect those changes.

Reported-by: auouymous <au@qzx.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>