libinotifytools: Rename init variable to fix conflict with entry point
Running inotifywait 3.22.1.0 on Ubuntu 22.04, I observed a bug
apparently due to a bad file descriptor:

    $ inotifywait -e modify foo
    Setting up watches.
    Couldn't watch foo: Bad file descriptor

I could also reproduce with the master branch. Investigating further, it
appears that inotifytools_init() in inotifytools.c exits immediately
after entry, because the global variable "init" is non-null, even though
we just called it at the beginning of inotifywait.c's main(), and the
global variable is supposed to be initialised at 0.

Bisecting showed that this variable used to be static, but was changed
at some point to non-static to share it with stats.c, and this is where
the bug fist occurs. As far as I understand, the symbol conflicts with
the init() function from the program's entry point [0]. When entering
inotifytools_init(), it seems that the value for that function is picked
up and compared to 0, found non-null, and the initialization function
exits, leaving inotify_fd uninitialized (-1) and a bad file descriptor.

Renaming the variable to something other than "init" is enough to fix
the bug.

    $ ./src/inotifywait -e modify foo
    Setting up watches.
    Watches established
    <waits>

[0] https://www.geeksforgeeks.org/executing-main-in-c-behind-the-scene/

Fixes: dd59c5347af8 ("Extract stats into their own file (#129)")