Fix fanotify_supported()
Parse negative value of --timeout argument and error only after setting
up all watches.

this is used to improve the helper to detect support for fanotify on a
specific filesystem (e.g. overlayfs) and test permission for specific
flags (e.g. --filesystem).

With this improvement, we can remove the workaround for not testing
fanotify inside containers, because the helper will detect if kernel
supports fanotify support inside containers or not (overlayfs support
for fanotify should be supported in future kernels).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>