Support watching a mount with fanotify
With fsnotify{wait,watch} --mount, try to setup a mount mark.
The --mount option is conflicting with options --filesystem and
both are conflicting with option --inotify.

Note that some events (e.g. create,delete) are not supported with
a mount mark, so this mode is limited to users that want the legacy
fanotify events (e.g. open,access,modify,close).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>