Suppress events with unknown path
When a filesystem watch is placed on a bind mount path, the events
that happen outisde of the bind mount path will be printed with empty
path, e.g.:
$ mkdir t
$ sudo mount --bind t t
$ sudo fanotifywait -S t &
$ touch x
// OPEN x

Skip events whose path cannot be resolved via /proc/self/fd symlink
such as events in paths not accessible from a bind mount which the
filesystem watch was set, to avoid printing events with empty path.

This can also be used as a feature of "subtree watch" by creating
a bind mount on the subtree root.

Add a test case to fsnotifywait-subtree.t where a file is created
outside of the bind mount path and event is ignored.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>